chore(engine): split builtin functions into different classes
