+++ pom.xml
@@ -226,6 +226,14 @@
             <!-- Clirr has problems with the inner classes that are generated by Scala (ending with "$") -->
             <exclude>org/camunda/feel/**/*$</exclude>
           </excludes>
+          <ignored>
+            <ignored>
+              <!-- ignore changes of the FEEL engine constructor, Java application should use the builder instead -->
+              <className>org/camunda/feel/FeelEngine</className>
+              <differenceType>7004</differenceType>
+              <method>FeelEngine(*)</method>
+            </ignored>
+          </ignored>
         </configuration>
       </plugin>
